refactor(videos): 라우팅 방식과 동일하게 폴더명 수정
 라우팅 방식과 동일하게 폴더명을
videoanalysis=>videos로 변경
