+++ src/entities/videos/model/types.ts
