+++ src/entities/videos/ui/VideoCard.tsx
