+++ src/features/videos/index.ts
@@ -1,0 +1,3 @@
+export { useVideos } from './model/useVideos'
+export { mockVideos } from './mock/mockVideos'
+export { VideoList } from './ui/VideoList'
