--- src/pages/videoAnalysis/index.ts
@@ -1,1 +1,0 @@
-export { VideoAnalysisPage } from './ui/VideoAnalysisPage'
