+++ src/pages/videos/index.ts
@@ -1,0 +1,1 @@
+export { VideosPage } from './ui/VideosPage'
