--- src/shared/api/msw/handlers/videoAnalysisHandlers.ts
@@ -1,16 +1,0 @@
-import { http, HttpResponse } from 'msw'
-
-import { mockVideoAnalysis } from '@/features/videoAnalysis/mock/mockVideoAnalysis'
-
-export const videoAnalysisHandlers = [
-  http.get(
-    `${process.env.NEXT_PUBLIC_API_URL}/channel/:id/videos`,
-    () => {
-      return HttpResponse.json({
-        isSuccess: true,
-        responseDto: mockVideoAnalysis,
-        error: null,
-      })
-    }
-  ),
-]
